refactor(chat-channel.component.ts): Moved chat channel to the chat directory and changed it into a standalone component
