+++ src/BurstChat.Client/src/app/components/chat/chat-channel/chat-channel.component.html
@@ -1,0 +1,16 @@
+@if(chatFound) {
+    <div class="h-100">
+        <burst-chat-info [options]="options"/>
+        <div class="channel-chat-content">
+            <burst-chat-messages [options]="options"/>
+        </div>
+        <burst-chat-input [options]="options"/>
+    </div>
+    <burst-message-edit-dialog [(visible)]="editMessageData.visible" [options]="options"
+        [message]="editMessageData.message"/>
+    <burst-message-delete-dialog [(visible)]="deleteMessageData.visible" [options]="options"
+        [message]="deleteMessageData.message"/>
+} @else {
+    <div></div>
+}
+
