+++ src/BurstChat.Client/src/app/components/chat/chat-channel/chat-channel.component.scss
@@ -1,5 +1,5 @@
-@import '../../../../../scss/theme';
-@import '../../../../../scss/sizing';
+@import '../../../../scss/theme';
+@import '../../../../scss/sizing';
 
 .channel-chat-content {
     @extend .h-100;
