+++ src/BurstChat.Client/src/app/components/chat/chat-channel/chat-channel.component.spec.ts
