--- src/BurstChat.Client/src/app/modules/chat/components/chat-channel/chat-channel.component.html
@@ -1,21 +1,0 @@
-<div class="h-100" *ngIf="!noChatFound">
-
-    <burst-chat-info [options]="options"></burst-chat-info>
-
-    <div class="channel-chat-content">
-
-        <burst-chat-messages [options]="options"></burst-chat-messages>
-
-    </div>
-
-    <burst-chat-input [options]="options"></burst-chat-input>
-
-</div>
-
-<burst-message-edit-dialog [(visible)]="editMessageData.visible" [options]="options"
-    [message]="editMessageData.message"></burst-message-edit-dialog>
-
-<burst-message-delete-dialog [(visible)]="deleteMessageData.visible" [options]="options"
-    [message]="deleteMessageData.message"></burst-message-delete-dialog>
-
-<div *ngIf="noChatFound"></div>
